test: rewrite test-child-process-spawn-args
- Use ESM for top-level await
- Avoid concurrent child processes in the test
- Use `events.once` instead of Promise wrapper
- Assert directly on sets
